Move API functional tests into `api`; organize
- Create new `annotations_test` and `api_test` test modules
