+++ tests/functional/api/test_annotations.py
@@ -14,7 +14,7 @@ native_str = str
 
 
 @pytest.mark.functional
-class TestAPI(object):
+class TestGetAnnotations(object):
     def test_api_index(self, app):
         """
         Test the API index view.
@@ -39,6 +39,8 @@ class TestAPI(object):
         assert data['@context'] == 'http://www.w3.org/ns/anno.jsonld'
         assert data['id'] == 'http://example.com/a/' + annotation.id
 
+
+class TestWriteAnnotation(object):
     def test_annotation_write_unauthorized_group(self, app, user_with_token, non_writeable_group):
         """
         Write an annotation to a group that doesn't allow writes.
@@ -66,24 +68,6 @@ class TestAPI(object):
 
         assert res.status_code == 400
         assert res.json['reason'].startswith('group:')
-
-    def test_cors_preflight(self, app):
-        # Simulate a CORS preflight request made by the browser from a client
-        # hosted on a domain other than the one the service is running on.
-        #
-        # Note that no `Authorization` header is set.
-        origin = 'https://custom-client.herokuapp.com'
-        headers = {'Access-Control-Request-Headers': str('authorization,content-type'),
-                   'Access-Control-Request-Method': str('POST'),
-                   'Origin': str(origin)}
-
-        res = app.options('/api/annotations', headers=headers)
-
-        assert res.status_code == 200
-        assert res.headers['Access-Control-Allow-Origin'] == str(origin)
-        assert 'POST' in res.headers['Access-Control-Allow-Methods']
-        for header in ['Authorization', 'Content-Type', 'X-Client-Id']:
-            assert header in res.headers['Access-Control-Allow-Headers']
 
 
 @pytest.fixture
@@ -115,13 +99,6 @@ def auth_client(db_session, factories):
     auth_client = factories.AuthClient(authority='thirdparty.example.org')
     db_session.commit()
     return auth_client
-
-
-@pytest.fixture
-def open_group(auth_client, db_session, factories):
-    group = factories.OpenGroup(authority=auth_client.authority)
-    db_session.commit()
-    return group
 
 
 @pytest.fixture
